requirements.txt
tensorboardX
tqdm
easydict
prettytable
chumpy
opencv-python
imageio-ffmpeg
matplotlib
roma
ipdb
pytorch-metric-learning                         # For one-shot action recognition
smplx[all]                                      # For mesh recovery
torch
torchvision
shap